pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neo-reels-backend"
version = "0.1.0"
description = "Short video backend MVP"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.29.0",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic-settings>=2.2.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.1.0",
  "boto3>=1.34.0",
  "celery>=5.3.6",
  "redis>=5.0.4",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = false
warn_redundant_casts = false
warn_unused_configs = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
